refactor: Render main contents using one component
