--- src/components/mainContents/BLContents.js
@@ -1,25 +1,0 @@
-import {
-  AppDownloadLink,
-  GenreTop,
-  MainNav,
-  CategoryMenu,
-} from '../index.js';
-import { Categories } from '../../data';
-import { convertStringToHTML } from '../../utils.js';
-
-const BLContents = () => {
-
-  const div = convertStringToHTML('<div class="main-contents"></div>');
-  const components = [
-    MainNav(),
-    CategoryMenu(Categories.bl),
-    GenreTop('BL/GL'),
-    AppDownloadLink(),
-  ];
-
-  div.append(...components);
-
-  return div;
-};
-
-export default BLContents;
