+++ src/data/components.js
@@ -1,16 +1,18 @@
 import {
   WebtoonPage,
   TempPage,
-  HomeContents,
-  DailyContents,
-  WebtoonContents,
-  ShonenContents,
-  DramaContents,
-  RomanceContents,
-  RomanceFantasyContents,
-  ActionContents,
-  BLContents,
+  MainNav,
+  CategoryMenu,
+  GenreTop,
+  AppDownloadLink,
+  MainBanner,
+  SubBanner,
+  DailySeriesRanking,
+  AnticipatedWork,
+  DailyRanking,
+  RecommendEvent,
 } from '../components';
+import { Categories } from './index.js';
 
 const PageComponent = {
   home: () => TempPage('Home'),
@@ -21,19 +23,74 @@ const PageComponent = {
   book: () => TempPage('Book'),
 };
 
-const ContentsComponent = {
-  home: () => HomeContents(),
-  daily: () => DailyContents(),
-  webtoon: () => WebtoonContents(),
-  shonen: () => ShonenContents(),
-  drama: () => DramaContents(),
-  romance: () => RomanceContents(),
-  romanceFantasy: () => RomanceFantasyContents(),
-  action: () => ActionContents(),
-  bl: () => BLContents(),
+const sectionComponent = {
+  action: () => [
+    MainNav(),
+    CategoryMenu(Categories.action),
+    GenreTop('액션무협'),
+    AppDownloadLink(),
+  ],
+  bl: () => [
+    MainNav(),
+    CategoryMenu(Categories.bl),
+    GenreTop('BL/GL'),
+    AppDownloadLink(),
+  ],
+  daily: () => [
+    MainNav(),
+    MainBanner(),
+    SubBanner(),
+    AppDownloadLink(),
+  ],
+  drama: () => [
+    MainNav(),
+    CategoryMenu(Categories.drama),
+    GenreTop('드라마'),
+    AppDownloadLink(),
+  ],
+  home: () => [
+    MainNav(),
+    MainBanner(),
+    CategoryMenu(Categories.home),
+    SubBanner(),
+    DailySeriesRanking(),
+    AnticipatedWork(),
+    GenreTop('로맨스'),
+    GenreTop('로판'),
+    GenreTop('드라마'),
+    GenreTop('BL/GL'),
+    GenreTop('소년'),
+    GenreTop('액션무협'),
+    DailyRanking(),
+    RecommendEvent(),
+    AppDownloadLink(),
+  ],
+  romance: () => [
+    MainNav(),
+    CategoryMenu(Categories.romance),
+    GenreTop('로맨스'),
+    AppDownloadLink(),
+  ],
+  romanceFantasy: () => [
+    MainNav(),
+    CategoryMenu(Categories.romanceFantasy),
+    GenreTop('로판'),
+    AppDownloadLink(),
+  ],
+  shonen: () => [
+    MainNav(),
+    CategoryMenu(Categories.shonen),
+    GenreTop('소년'),
+    AppDownloadLink(),
+  ],
+  webtoon: () => [
+    MainNav(),
+    DailyRanking(),
+    AppDownloadLink(),
+  ],
 };
 
 export {
   PageComponent,
-  ContentsComponent,
+  sectionComponent,
 };
